+++ packages/react/src/lib/accessibility/Behaviors/Toolbar/menuAsToolbarBehavior.ts
@@ -1,18 +1,18 @@
 import { Accessibility, FocusZoneMode } from '../../types'
 import { FocusZoneDirection } from '../../FocusZone'
-import toolbarButtonBehavior from './toolbarButtonBehavior'
+import menuItemAsToolbarButtonBehavior from './menuItemAsToolbarButtonBehavior'
 
 /**
  * @description
  * Implements ARIA Toolbar design pattern.
- * Child item components need to have toolbarButtonBehavior assigned.
+ * Child item components need to have menuItemAsToolbarButtonBehavior assigned.
  * @specification
  * Adds role 'toolbar' to 'root' component's part.
  * Embeds component into FocusZone.
  * Provides arrow key navigation in bidirectional direction.
  * When component's container element receives focus, focus will be set to the default focusable child element of the component.
  */
-const toolbarBehavior: Accessibility = (props: any) => ({
+const menuAsToolbarBehavior: Accessibility = (props: any) => ({
   attributes: {
     root: {
       role: 'toolbar',
@@ -26,8 +26,8 @@ const toolbarBehavior: Accessibility = (props: any) => ({
     },
   },
   childBehaviors: {
-    item: toolbarButtonBehavior,
+    item: menuItemAsToolbarButtonBehavior,
   },
 })
 
-export default toolbarBehavior
+export default menuAsToolbarBehavior