+++ packages/react/src/lib/positioner/Popper.tsx
@@ -22,7 +22,7 @@ const Popper: React.FunctionComponent<PopperProps> = props => {
   const {
     align,
     children,
-    eventsEnabled,
+    enabled,
     modifiers: userModifiers,
     offset,
     pointerTargetRef,
@@ -51,14 +51,33 @@ const Popper: React.FunctionComponent<PopperProps> = props => {
     [rtl, offset, position],
   )
 
-  React.useEffect(
+  const scheduleUpdate = React.useCallback(() => {
+    if (popperRef.current) {
+      popperRef.current.scheduleUpdate()
+    }
+  }, [])
+
+  const destroyInstance = React.useCallback(() => {
+    if (popperRef.current) {
+      popperRef.current.destroy()
+      popperRef.current = null
+    }
+  }, [])
+
+  const createInstance = React.useCallback(
     () => {
+      destroyInstance()
+
+      if (!enabled || !targetRef.current || !contentRef.current) {
+        return
+      }
+
       const pointerTargetRefElement = pointerTargetRef && pointerTargetRef.current
       const popperHasScrollableParent = getScrollParent(contentRef.current) !== document.body
 
       const modifiers: PopperJS.Modifiers = _.merge(
         { preventOverflow: { padding: 0 } },
-        { flip: { padding: 0 } },
+        { flip: { padding: 0, flipVariationsByContent: true } },
         /**
          * When the popper box is placed in the context of a scrollable element, we need to set
          * preventOverflow.escapeWithReference to true and flip.boundariesElement to 'scrollParent' (default is 'viewport')
@@ -92,7 +111,6 @@ const Popper: React.FunctionComponent<PopperProps> = props => {
 
       const options: PopperJS.PopperOptions = {
         placement: proposedPlacement,
-        eventsEnabled,
         positionFixed,
         modifiers,
         onCreate: handleUpdate,
@@ -100,31 +118,44 @@ const Popper: React.FunctionComponent<PopperProps> = props => {
       }
 
       popperRef.current = createPopper(targetRef.current, contentRef.current, options)
-      return () => popperRef.current.destroy()
     },
-    [computedModifiers, eventsEnabled, userModifiers, positionFixed, proposedPlacement],
+    // TODO review dependencies for popperHasScrollableParent
+    [computedModifiers, enabled, userModifiers, positionFixed, proposedPlacement],
   )
 
   React.useEffect(
     () => {
-      popperRef.current.scheduleUpdate()
+      createInstance()
+      return destroyInstance
     },
-    [...positioningDependencies, computedPlacement],
+    [createInstance],
   )
+
+  React.useEffect(scheduleUpdate, [...positioningDependencies, computedPlacement])
 
   const child =
     typeof children === 'function'
       ? (children as PopperChildrenFn)({
           placement: computedPlacement,
-          scheduleUpdate: () => popperRef.current && popperRef.current.scheduleUpdate(),
+          scheduleUpdate,
         })
-      : React.Children.only(children)
+      : children
 
-  return <Ref innerRef={contentRef}>{child as React.ReactElement}</Ref>
+  return (
+    <Ref
+      innerRef={contentElement => {
+        contentRef.current = contentElement
+        // for correct positioning we need to create the PopperJS instance immediately after we get a ref to the popper box
+        createInstance()
+      }}
+    >
+      {React.Children.only(child) as React.ReactElement}
+    </Ref>
+  )
 }
 
 Popper.defaultProps = {
-  eventsEnabled: true,
+  enabled: true,
   positionFixed: false,
   positioningDependencies: [],
 }